Squash to: Cleanup synth_mode
* Handle '*' for freq
* Correct ext_lo_sb to int
* Correct input_sb to int
* Remove extraneous memcpy

+++ include/dbbc3_synthesizer_ds.h
@@ -38,11 +38,11 @@ struct dbbc3_synthesizer_cmd {
     struct m5state state;
   } ext_lo_freq;
   struct {
-    double ext_lo_sb;
+    int ext_lo_sb;
     struct m5state state;
   } ext_lo_sb;
   struct {
-    double input_sb;
+    int input_sb;
     struct m5state state;
   } input_sb;
 };
